Add standalone gallery page
With the full list of company screenshots. This involved refactoring the
gallery to a single include shared by the homepage and dedicated page.

We just have to add the extended css class to any new companies that we
want to show up on the dedicated page rather than the homepage.